Feat (UI): Many changes.
Activated the OnScrollListener on Jobs Recycler View.
Activated the OnClickListeners on Jobs Recycler View.
Starting the design for Job Detail View.
Added Ken Burns effect on background.
Added a random background selection for Ken Burns effect.
Added a touchable button to go the main website to see the proposal from the source.
Fixed other problems with the UI.

+++ app/src/main/java/com/cyberfanta/talentviewer/models/Bios.kt
@@ -296,7 +296,7 @@ data class Person2(
 	val flags: Flags2? = null,
 
 	@field:SerializedName("weight")
-	val weight: Int? = null,
+	val weight: Float? = null,
 
 	@field:SerializedName("locale")
 	val locale: String? = null,
@@ -629,7 +629,7 @@ data class StrengthsItem2(
 	val name: String? = null,
 
 	@field:SerializedName("weight")
-	val weight: Int? = null,
+	val weight: Float? = null,
 
 	@field:SerializedName("id")
 	val id: String? = null,
